requirements.txt
opencv-python>=4.5.5.62
numpy
pillow>=10.1.0
scipy>=1.12.0
litelama>=0.1.7
